Rewrite ff8_data.h
